[tooltip] Hacking retarded event customization
This change allows you to pass in 0 events and manage the hide/show of the tooltip yourself. This is useful in certain cases when you have to hide/show events based on some logic, and not the usual mouseover/mouseout.

I needed this fix so I could get the tooltip working with hoverIntent.