[expose] Adding deferLoad behavior (default false), and fixing stuff
Major update to expose that makes it much easier to work with. By default now, load() is called automatically on exposing. It will handle initializing the elements (if not already initialized). Use deferLoad: true if you want to wait till load() is called to set things up.

Also added exposedClass which will turn on and off when exposed and not exposed.